worker: stop doing heavy work on the box that runs agent turns
Three worker deaths on 2026-07-30, one root cause. An agent turn runs LOCALLY
on the dispatcher by design — it is network-bound waiting on an LLM — and three
different tools do real compute inside one, on the smallest box in the fleet.
An OOM kill is a disappearance, not an exception, so every one of them reached
the customer as "I lost my connection while working on that request", and
MAX_ATTEMPTS_AGENT is 1, so none of them retried.

record_website: the FIRST production call of this feature killed the worker.
A customer asked for Google Search b-roll at 13:15:28; the model asked for
three captures at 13:15:57; last heartbeat 13:16:15. A full headless Chromium
at 1080x1920 with a CDP screencast draining frames. The executor already has
Chromium (one image, two roles), 8 vCPU and 32Gi, so the browser runs there
and only the storage key comes back — the mp4 never crosses the wire. A remote
failure deliberately does NOT fall back to the local browser: that would
reproduce the crash on the box we know is too small, on the retry, when the
user has waited longest.

add_text_behind: killed the worker one second after being called, on "put the
title behind me walking". The plate comment says 24 samples is "~37 MB" and it
was right — the code widened every sample to float32 on the way in, 4x, then
np.stack copied the lot and np.median partitioned another copy. Measured on
the real shape: 312.6 MB peak, against 87.2 MB in uint8, with a byte-identical
plate (median is order statistics; the dtype it partitions in cannot change
which value wins).

set_volume: "remove the sound of the video" is close to the simplest request
this product gets, and the obvious way to say it is a very negative gain. -100
dB got "outside [-60.0, 12.0]" — a rejection for asking for more of exactly
what the parameter does. -60 dB is already inaudible, so below the floor now
clamps. Above the ceiling still refuses: too loud is a real defect.

Executor redeploy required — it gains the capture runner.

+++ worker/matte.py
@@ -119,10 +119,26 @@ def _plate(path, start, dur, w, h, extra_vf=None):
     for i in range(PLATE_SAMPLES):
         t = start + step * (i + 0.5)
         for f in _decode(path, t, 0.06, w, h, extra_vf):
-            stack.append(f.astype(np.float32))
+            # uint8, NOT float32 — see below. np.median partitions in the
+            # input dtype and returns float64 for the (h, w, 3) result, which
+            # is one frame's worth, not the stack's.
+            stack.append(f)
             break
     if len(stack) < 6:
         return None
+    # THE COMMENT ON PLATE_SAMPLES WAS RIGHT; THE CODE WAS NOT (round 61).
+    #
+    # "24 x 960x540x3 is ~37 MB" is the uint8 figure. Every sample was being
+    # widened to float32 on the way in — 4x — and then np.stack COPIED the lot
+    # and np.median partitioned another copy, so a 6-second title cost roughly
+    # 450 MB of peak RSS. This runs inside the agent turn, which runs on the
+    # dispatcher, and on 2026-07-30 it killed that process one second after
+    # add_text_behind was called: the user asked for a title behind themselves
+    # walking and got "I lost my connection while working on that request".
+    #
+    # Kept in uint8 the whole way, the same three allocations come to ~110 MB
+    # and the arithmetic is identical — median is order statistics, so the
+    # dtype it partitions in cannot change which value wins.
     return np.median(np.stack(stack), axis=0)
